requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
python-dotenv==1.0.0
openai==0.28.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.5.0
scipy==1.11.1
joblib==1.3.2
PyJWT==2.8.0